removed exceptions package and moved RaceTrackEndException into the racetrack package
