--- lib/src/main/java/com/group3/racingbot/exceptions/package-info.java
@@ -1,7 +1,0 @@
-
-/**
- * Contains exceptions for 
- * @author Maciej Bregisz
- *
- */
-package com.group3.racingbot.exceptions;
