+++ lib/src/main/java/com/group3/racingbot/racetrack/RaceTrackEndException.java
@@ -1,4 +1,4 @@
-package com.group3.racingbot.exceptions;
+package com.group3.racingbot.racetrack;
 
 
 /**
